refactor: todos os enums foram adicionados num pacote de enums dentro do domain, e todas as classes e interfaces que os importavam foram atualizadas.
