+++ back-end/src/main/java/br/edu/com/tg/manager/core/domain/entities/Student.java
@@ -1,6 +1,6 @@
 package br.edu.com.tg.manager.core.domain.entities;
 
-import br.edu.com.tg.manager.core.domain.entities.enums.StudentStatus;
+import br.edu.com.tg.manager.core.domain.enums.StudentStatus;
 import br.edu.com.tg.manager.core.domain.exceptions.DomainException;
 
 /**
